requirements.txt
torch
torchvision
numpy
pyyaml
matplotlib